requirements.txt
# Core Dependencies
openai-agents-sdk>=0.1.0,<1.0.0  # OpenAI Agents SDK for agent creation and orchestration
fastapi>=0.109.0,<1.0.0  # FastAPI for REST API implementation
uvicorn>=0.27.0,<1.0.0  # ASGI server for FastAPI
pydantic>=2.5.0,<3.0.0  # Data validation and settings management

# Database and Caching
sqlalchemy>=2.0.0,<3.0.0  # ORM for database access
alembic>=1.13.0,<2.0.0  # Database migration tool
asyncpg>=0.29.0,<1.0.0  # PostgreSQL async driver
psycopg2-binary>=2.9.0,<3.0.0  # PostgreSQL driver (binary version)
redis>=5.0.0,<6.0.0  # Redis client for caching
databases>=0.7.0,<1.0.0  # Async database support

# API and HTTP
httpx>=0.26.0,<1.0.0  # Async HTTP client (for SERP API calls)
python-multipart>=0.0.6,<1.0.0  # For handling form data in FastAPI
python-dotenv>=1.0.0,<2.0.0  # Environment variable management
starlette>=0.36.0,<1.0.0  # ASGI framework (used by FastAPI)
email-validator>=2.1.0,<3.0.0  # Email validation for Pydantic
orjson>=3.9.0,<4.0.0  # Fast JSON serialization/deserialization

# Security
python-jose>=3.3.0,<4.0.0  # JWT token handling for API auth
passlib>=1.7.4,<2.0.0  # Password hashing
bcrypt>=4.1.0,<5.0.0  # Password hashing algorithm
cryptography>=42.0.0,<43.0.0  # Cryptography libraries

# Data Processing
numpy>=1.24.0,<2.0.0  # Numerical Python
pandas>=2.0.0,<3.0.0  # Data analysis and manipulation
nltk>=3.8.0,<4.0.0  # Natural language toolkit for keyword extraction
scikit-learn>=1.3.0,<2.0.0  # Machine learning utilities for text classification
spacy>=3.7.0,<4.0.0  # NLP for keyword extraction and intent detection

# Observability
loguru>=0.7.0,<1.0.0  # Structured logging
opentelemetry-api>=1.22.0,<2.0.0  # OpenTelemetry API
opentelemetry-sdk>=1.22.0,<2.0.0  # OpenTelemetry SDK
opentelemetry-exporter-otlp>=1.22.0,<2.0.0  # OpenTelemetry OTLP exporter
prometheus-client>=0.19.0,<1.0.0  # Prometheus metrics

# Testing
pytest>=8.0.0,<9.0.0  # Testing framework
pytest-asyncio>=0.23.0,<1.0.0  # Async testing support
pytest-cov>=4.1.0,<5.0.0  # Test coverage
requests-mock>=1.11.0,<2.0.0  # Mock HTTP requests
fakeredis>=2.20.0,<3.0.0  # Mock Redis for testing
freezegun>=1.3.0,<2.0.0  # Time freezing for tests

# Development
black>=24.1.0,<25.0.0  # Code formatter
isort>=5.13.0,<6.0.0  # Import sorter
flake8>=7.0.0,<8.0.0  # Linter
mypy>=1.8.0,<2.0.0  # Static type checking
pre-commit>=3.6.0,<4.0.0  # Pre-commit hooks

# Documentation
mkdocs>=1.5.0,<2.0.0  # Documentation generator
mkdocs-material>=9.5.0,<10.0.0  # Material theme for MkDocs
mkdocstrings>=0.24.0,<1.0.0  # API documentation generator

# API Client (Optional SERP API choices)
serpapi-python>=0.1.0,<1.0.0  # SerpAPI client (optional)
# Or another SERP API client of your choice

# Performance metrics
psutil>=5.9.0
prometheus-client>=0.14.1 